fix: install to temp directory using --tree instead of --local
@teto

This also adds an `upload` input that allows us to
test the first install step on each PR, and hopefully prevent
future breakage.

I am keeping the input private so that we can remove it again if
we find a way to do temporary test uploads to LuaRocks, as
discussed in #26.